% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_flextable.R
\name{fct_create_flextable}
\alias{fct_create_flextable}
\title{Generate a flextable from the pretty printed version of the xlsform}
\usage{
fct_create_flextable(xlsformpath, label_language, logic)
}
\arguments{
\item{xlsformpath}{path to the file with xlsform}

\item{label_language}{Language to be used in case you have more than one. 
If not specified, the 'default_language' in the 'settings' worksheet is used.
 If that is not specified and more than one language is in the XlsForm, 
 the language that comes first within column order will be used.}

\item{logic}{TRUE per default put to FALSE to take out the printing of 
constraints.}
}
\value{
a flextable ready to be printed in an Rmd template for word - 
         using officedown.
}
\description{
Note that better legibility, it is advised to put the select 
question with many possible answers - typically something like 
"what is your country of origin?" -  as 'select_from_file'
}
\examples{
fct_create_flextable(
  xlsformpath = system.file("demo.xlsx", package = "XlsFormUtil"),
  label_language = NULL )
}
