% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_render_prettyprint.R
\name{fct_render_prettyprint}
\alias{fct_render_prettyprint}
\title{Render Word version of an xlsform}
\usage{
fct_render_prettyprint(
  dir,
  xlsformfile,
  label_language = NULL,
  logic = TRUE,
  estimation = TRUE,
  wpm = 180,
  maxmodalities = 7,
  resptimeclose = 4,
  resptimecondopen = 7,
  resptimeopen = 10,
  avrgrepeat = 3
)
}
\arguments{
\item{dir}{directory where the xlsform is located}

\item{xlsformfile}{name of the file for the xlsform}

\item{label_language}{language to sue to generate the pretty print -
for instance _"Español (es)"_, _"Français (fr)"_ or _"English (en)"_ .
 If not specified, the 'default_language' in the 'settings' worksheet 
 is used. If that is not specified and more than one language is in
  the XlsForm, the language that comes first within column order will
   be used.}

\item{logic}{TRUE per default put to FALSE to take out the printing of 
constraints.}

\item{estimation}{TRUE FALSE add an estimation of interview dration}

\item{wpm}{word per minute - an average 180 word per minute (per default) 
required to read loudly the text}

\item{maxmodalities}{if more than 7 potential answers for a select question 
(per default)- then we assume that those modalities will not be read by the
 enumerator - but rather selected based on an open answer - and not be 
 accounted for the modalities duration estimation}

\item{resptimeclose}{an average 4 seconds (per default) for respondent
to reply for closed questions}

\item{resptimecondopen}{an average of  7 seconds (per default) to reply 
to conditional text question (accounting for question type of 
"other, please specify").}

\item{resptimeopen}{an average of  10 seconds (per default) to reply 
to open text question.}

\item{avrgrepeat}{In case of repeat questions, an average 3 repeat 
(per default) is accounted for.}
}
\value{
a word document
}
\description{
Us the built in template and functions from the package to generate a word 
version of the xlsform
}
\details{
When customizing and adjusting a household survey questionnaire during the 
design phase, it's often necessary to have one testing version 
(i.e. encoded in [xlsform](http://xlsform)) and a more legible version in word
 that can be then shared with non-technical experts for them to comment and review.
 
 Moving between a paper version and an encoded-machine ready version is not 
 smooth. Instead of having the master version in word and updating once while 
 the xlsform, it's more convenient to generate a word output to collect 
 feedback in word tracking mode.
 
 The default _prettyprinting_ template brings in a single grid different parts of the xlsform. The _prettyprinting_ can be done separately for each different language in case the `xlsform` includes more than one. The template increases legibility for both __question blocks__:
 
   1. begin_group --> depending on the level are output with a different style of heading level
   2. begin_repeat --> displayed as a header but with a specific distinct color and note
 
 and within each block, each __question details__ is  included:  
 
   1. question code & type
   2. question label & hint
   3. if select_one or select_multiple, modalities code and label as a nested table
   4. if present - question constraint with warning message as well as related question skip logic
}
\examples{
# fct_render_prettyprint( dir =  system.file("", package = "XlsFormUtil"),
#   xlsformfile = "demo.xlsx",
#   label_language = NULL )
}
