% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tabulate_form.R
\name{fct_tabulate_form}
\alias{fct_tabulate_form}
\title{Combine the xlsform in single table for pretty printing}
\usage{
fct_tabulate_form(xlsformpath, label_language = NULL, logic = TRUE)
}
\arguments{
\item{xlsformpath}{path to the file with xlsform}

\item{label_language}{Language to be used in case you have more than one. 
If not specified, the 'default_language' in the 'settings' worksheet is used.
 If that is not specified and more than one language is in the XlsForm, 
 the language that comes first within column order will be used.}

\item{logic}{TRUE per default put to FALSE to take out the printing of 
constraints.}
}
\value{
a single data frame
}
\description{
Combine the xlsform in single table for pretty printing
}
\examples{
# fct_tabulate_form()

prettyform <- fct_tabulate_form( xlsformpath = system.file("demo.xlsx", 
                                              package = "XlsFormUtil") )

knitr::kable(utils::head(prettyform, 10))

}
