% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_xlsfrom_language.R
\name{fct_xlsfrom_language}
\alias{fct_xlsfrom_language}
\title{fct_xlsfrom_language}
\usage{
fct_xlsfrom_language(xlsformpath)
}
\arguments{
\item{xlsformpath}{path to the file with xlsform}
}
\value{
label_language
}
\description{
Retrieve a factor with all label languages used in a specific xlsform.
Used in the app to retrieve the language to consider based on the form
}
\examples{
## With different language
fct_xlsfrom_language( xlsformpath = system.file("demo.xlsx",
                                                package = "XlsFormUtil"))

## without language
fct_xlsfrom_language( xlsformpath = system.file("KoBoXLSFormExamples.xlsx",
                                                package = "XlsFormUtil"))
}
